# colexec: speed up ordered sync