--- a/pkg/sql/colexec/execgen/cmd/execgen/orderedsynchronizer_gen.go
+++ b/pkg/sql/colexec/execgen/cmd/execgen/orderedsynchronizer_gen.go
@@ -1,0 +1,46 @@
+// Copyright 2020 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package main
+
+import (
+	"io"
+	"io/ioutil"
+	"strings"
+	"text/template"
+
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+)
+
+func genOrderedSynchronizer(wr io.Writer) error {
+	d, err := ioutil.ReadFile("pkg/sql/colexec/orderedsynchronizer_tmpl.go")
+	if err != nil {
+		return err
+	}
+
+	s := string(d)
+
+	// Replace the template variables.
+	s = strings.Replace(s, "_TYPE", "coltypes.{{.LTyp}}", -1)
+	s = strings.Replace(s, "_TemplateType", "{{.LTyp}}", -1)
+
+	s = replaceManipulationFuncs(".LTyp", s)
+
+	tmpl, err := template.New("orderedsynchronizer").Parse(s)
+	if err != nil {
+		return err
+	}
+
+	return tmpl.Execute(wr, sameTypeComparisonOpToOverloads[tree.EQ])
+}
+
+func init() {
+	registerGenerator(genOrderedSynchronizer, "orderedsynchronizer.eg.go")
+}