**colexec: refactor ordered sync to use assign instead of append**

Ordered synchronizer sets the output one row at a time. Previously, we
were calling Append on the output vectors, but it is faster to use Sets
in such scenario.

Release note: None

**colexec: speed up ordered sync**

This commit speeds up the ordered synchronizer with two modifications:
1. instead of updating the allocator after we copy a single whole row,
we will now update the allocator after we copy a full batch of rows
2. to reduce the number of interface conversions we will "strip" output
vectors to get to the underlying slices and will use those directly when
copying the data.

Release note: None