@yuzefovich 2020-01-08
Here is the benchmark with the default three inputs:
```
name                    old time/op    new time/op    delta
OrderedSynchronizer-16     139µs ± 1%      55µs ± 1%   -60.54%  (p=0.000 n=9+10)

name                    old speed      new speed      delta
OrderedSynchronizer-16   177MB/s ± 1%   449MB/s ± 1%  +153.45%  (p=0.000 n=9+10)

name                    old alloc/op   new alloc/op   delta
OrderedSynchronizer-16    32.8kB ± 0%     0.0kB       -100.00%  (p=0.000 n=10+10)

name                    old allocs/op  new allocs/op  delta
OrderedSynchronizer-16     1.02k ± 0%     0.00k       -100.00%  (p=0.000 n=10+10)
```