@yuzefovich 2020-01-09
There is currently a bug with Setting values on Bytes type, I'll fix it tomorrow.