@yuzefovich 2020-01-09
Actually, I was wrong - there is no bug in this PR, so it is RFAL.

(Yesterday I thought that when we're setting NULL value on a `Bytes` vector, we would need to set an empty byte slice in the corresponding position as well, but this is not necessary because before returning the output batch, we're setting a length on it, and that method will enforce the invariant on `Bytes`.)