@yuzefovich 2020-01-09
I think the most of the performance improvement comes from reducing the number of calls on allocator, and I was the one who put them in place :)

TFTR!

bors r+